orbit: acquire all the CelestialBody orbital parameters from config
This will be really helpful for patching SOIs.
Also we could probably simplify some of our existing RocketSim3D things
to use this information rather than asking Telemachus.
However, as previously noted, we get the wrong mean motion when we try
to work it out ourselves; need to figure out why, and fix that, first.