Filter bad data in ThrottleGauge, TWRGauge and Light
If they get a value of 2, that means Telemachus decided not to give us
data (it apparently means 'no power', and happens even though we have
plenty of power...) so they should treat it the same as getting no data.
Also, made ThrottleGauge display missing data with the standard yellow-
background "NO DATA", instead of displaying "N/A" (which is only desired
for FuelGauge, which happens to share PercentageGauge).